backend/requirements.txt
fastapi
uvicorn[standard]
anthropic>=0.40.0
firecrawl-py>=1.0.0
tavily-python>=0.5.0
PyGithub>=2.0.0
resend>=2.0.0
playwright>=1.40.0
Pillow>=10.0.0
Jinja2>=3.1.0
pydantic>=2.0.0
httpx>=0.27.0
APScheduler>=3.10.0
